# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KKN Integration Dashboard API",
        "description": "Dokumentasi API dashboard KKN. API ini dipakai untuk integrasi dengan aplikasi project-akhir (portal mahasiswa & dosen) yang mengirim data pendaftaran, dokumen, dan nilai ke dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Server lokal dashboard"
        }
    ],
    "paths": {
...
        "/api/luaran/{id}": {
            "get": {
                "tags": [
                    "Luaran"
                ],
                "summary": "Detail luaran",
                "description": "Mengambil detail satu luaran berdasarkan ID.",
                "operationId": "8638604d7d0bd23fe73ed426f293e928",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID luaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail luaran berhasil diambil"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
        "/api/luaran/{id}/status": {
            "put": {
                "tags": [
                    "Luaran"
                ],
                "summary": "Verifikasi / ubah status luaran",
                "description": "Mengubah status verifikasi luaran (pending, approved, rejected) beserta catatan opsional.",
                "operationId": "554b977762a85b58e4ba7773337f1bbe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID luaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "approved",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "rejected"
                                        ]
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Luaran sudah sesuai",
                                        "nullable": true,
                                        "maxLength": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status luaran berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/luaran/store-from-external": {
            "post": {
                "tags": [
                    "Luaran"
                ],
                "summary": "Menerima luaran dari project-akhir",
                "description": "Menyimpan luaran berupa file artikel (PDF/DOC/DOCX), link video aftermovie, dan link artikel. Endpoint ini juga tersedia sebagai alias di /api/luaran/store.",
                "operationId": "e449e392965674c636374f18a671829f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul_kegiatan",
                                    "video_aftermovie",
                                    "artikel_link",
                                    "artikel_file",
                                    "user_nim"
                                ],
                                "properties": {
                                    "judul_kegiatan": {
                                        "type": "string",
                                        "example": "KKN Tematik Desa Sukamaju",
                                        "maxLength": 255
                                    },
                                    "video_aftermovie": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://youtube.com/watch?v=xxxx"
                                    },
                                    "artikel_link": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://media.com/artikel-kkn"
                                    },
                                    "artikel_file": {
                                        "description": "File artikel PDF/DOC/DOCX (maks 10MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_nim": {
                                        "type": "string",
                                        "example": "2010001",
                                        "maxLength": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Luaran berhasil disimpan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Terjadi kesalahan server"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API key mitra di sini untuk endpoint tarik data nilai.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}